container and component PDP
